requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
Flask==2.3.3
razorpay==1.2.1
gunicorn==20.1.0
